1.1.Hello API Security + API Docs on Production?
1.Hello API Security + API Docs on Production?

Hiding the Docs on Production?

Head back to /api to see our beautiful, interactive, Swagger documentation. I know some of you are probably thinking: I know, this is great for development, but how can I disable this for production?

The answer is... you shouldn't!

Well first, let me show you how you can disable this on production and then I'll try to convince you to keep it.

Open up config/packages/api_platform.yaml. API Platform is highly configurable. So, reminder time: if you go to an open terminal, you can see all your current configuration - including any default values - by running:

php bin/console debug:config api_platform

You can also run:

php bin/console config:dump api_platform to see an example tree of all the possible keys with some explanations. One of the options is called enable_docs. Copy this. The goal is to disable the docs only in production, but let's start by disabling them everywhere to see how it works.

Set enable_docs to false.

Ok, go back and refresh /api. Um... there is absolutely no change! That's due to a small dev-only bug with a few of these options: when we change those options, they're not causing the routing cache to rebuild automatically. No big problem, clear that cache manually with:

php bin/console cache:clear

Fly back over, refresh and... yep! The documentation is gone. Oh... but instead of a 404, it's a 500 error!

It turns out that going to /api to see the docs was just a convenience. The documentation was really stored at /api/docs and this is now a 404. You could also go to /api/docs.json or /api/docs.jsonld before, but now it's all gone.

One of the unfortunate things about removing the documentation is that it wasn't just rendered as HTML. In part 1 of this series, we talked about JSON-LD and Hydra, and how API Platform generates machine-readable documentation to explain your API. If I go to /api/cheeses.jsonld, this advertises that we can go to /api/contexts/cheeses to get more "context", more machine-readable "meaning". Whelp, that doesn't work anymore. The point is: if you disable documentation, realize that you're also disabling the machine-friendly documentation.

And if you want to fix the 500 on /api, you also need to disable the "entrypoint". Right now, if we go to /api/index.jsonld, we get a, sort of, "homepage" for our API, which tells us what URLs we could go to next to discover more. When we go to /api, that's the HTML entrypoint and that's... totally broken.