1.3.Hello API Security + API Docs on Production?
So to fully disable your docs without a 500 error, you need both keys. To make this only happen in production, copy them, delete them, and, in the config/packages/prod directory, create a new api_platform.yaml file. Inside, start with api_platform: then paste.

If we changed to the prod environment and rebuilt the cache, we'd be done!

But instead... I'm going to comment this out... for two reasons. First, I like the documentation, I like the machine-readable documentation and I like the entrypoint: the documentation homepage. And second, more importantly, if you want to hide your documentation so that nobody will use your API, that's a bad plan. That's security through obscurity. If your API lives out on the web, you need to assume people will find it and you need to properly secure it. Hey! That's the topic of this tutorial!

Rebuild the cache one more time.

At the end of the day, if you're ok keeping the machine-readable docs and the entrypoint stuff, but you really don't want to expose the pretty HTML docs, you could always create an event subscriber on the kernel.request event - now called the RequestEvent in Symfony 4.3 - and, if the URL is /api and the request format is HTML, return a 404.

Ok, let's get into the real action: let's start figuring out how we're going to let users of our API log in.