7.1.Logout & Passing API Data to JS on Page Load{Adding Logout}
Hey! We can log in and our JavaScript even knows when we log in, and who we are: it prints our username & a log out link that goes to /logout... which doesn't actually work yet... cause we haven't enabled that in Symfony.

Adding Logout

Wait... but what does "logging out" even mean in an API context? Whelp, like everything, it depends on how you authenticate. Because we're using a session cookie, logging out basically means removing the user information from the session. If you were using some sort of API token, it would mean invalidating that token on your authentication server - like, removing it from some database table for tokens... again, it depends on your setup. We'll talk more about that type of authentication a bit later - on a special security part 2 of this tutorial.

Anyways, no surprise that Symfony has built-in support for logging the user out of the session. In config/packages/security.yaml, under our firewall add logout: then, below, say path: app_logout. Just like app_login, this is the name of a route that we're going to create next. When a user accesses this route, they'll be logged out.

To create that, open src/Controller/SecurityController.php and add public function logout() with @route() above. Set the URL to /logout and name it app_logout.

Just like with the app_login route, the route just... needs to exist... otherwise the user will see a 404 when they go to /logout. As long as it does exist, when the user goes to /logout, the logout mechanism will intercept the request, remove the user from the session, then redirect them to the homepage... which is configurable.

This means that, unless we've messed something up, the controller will never be reached. Let's scream in case it somehow is executed: Throw an exception with:

  should not be reached

Let's try the flow: move over, hit log out and... before it loads, you can see that we're currently logged in. And now... gone! We are anonymous.

Before we keep going with all this API & security goodness, our app has a bug. If we log in... as soon as the AJAX call finishes, we've made our Vue.js frontend smart enough to update and say that we're logged in. But when we refresh, that's gone! Gasp! Our web debug toolbar knows we're logged in... but our JavaScript does not. And... it makes sense: when we first load the page... if you look at the HTML source - you can ignore all the web debug toolbar stuff down here... the entire application looks like this. There's no HTML or JavaScript data that hints to Vue that we're authenticated.