7.2.Logout & Passing API Data to JS on Page Load{Passing Data to JavaScript on Page Load}
Shortcoming of previous commit

Before we keep going with all this API & security goodness, our app had a bug. If we log in... as soon as the AJAX call finished, we've made our Vue.js frontend smart enough to update and say that we're logged in. But when we refreshed, that's gone! Gasp! Our web debug toolbar knew we're logged in... but our JavaScript did not. And... it made sense: when we first loaded the page... if you look at the HTML source - you can ignore all the web debug toolbar stuff down here... the entire application looked like this. There's no HTML or JavaScript data that hints to Vue that we're authenticated.

Passing Data to JavaScript on Page Load

How can fix the problem described above?  First, as soon as the page loads, we could make an AJAX request to some endpoint and say:

  Hey! Who am I logged in as? Cause... I forgot?

To make that happen, we would need some sort of a /me endpoint: something that would return information about who we are. A useful, but not-so-RESTful endpoint.

The other option is quite nice: send the user data from your HTML into Vue on page load.

Open up the template for this page: templates/frontend/homepage.html.twig. Yep, nothing here but some small HTML to bootstrap the Vue app... though I am doing one interesting thing: I'm passing a prop to Vue called entrypoint. I'm not using this anywhere... but it's a cool example: entrypoint is the URL to our documentation homepage. In theory, we could use that dynamic URL in our Vue app to figure out what other URLs we could call... we would use our API like a browser: surfing through links. Anyways, this shows a nice way to pass simple data into Vue.

And.. we could pass the current user's IRI as another prop. Inside Vue, we would then make an AJAX call to that URL to get the user data... so, no need for the /me endpoint. That's really the simplest option, though it does have a minor downside: there will be a slight delay on page load before our app knows who's logged in.

Serializing Data Directly to JavaScript

To avoid that AJAX call, we can dump that data directly into Vue. Check it out: start in src/Controller/FrontendController.php. This is the controller behind the homepage.html.twig template.. and it's not super impressive. Add a SerializerInterface $serializer argument... and then pass a new variable to the template called user. I want this to be the JSON-LD version of our user - the exact same thing we would get from making an AJAX request. Set this to $serializer->serialize() passing it $this->getUser() and jsonld for the format. If the user is not logged in, the new user variable will be null... but if they are logged in, we'll get our big, nice JSON-LD structure.

Now that we have this, create a script tag and set the data on a global variable... how about: window.user = {{ user|raw }}.

Hey! Our user data is accessible to JavaScript!

Head over to CheeseWhizApp to use it. Generally speaking, I try not to use or reference global variables from my JavaScript. As a compromise, I like to only reference global variables from my top level component. If a deeper component needs it, I'll pass it down as a prop.

Create a new mounted() function - Vue will automatically call this after the component is "mounted" to the page - and if window.user, so, if it's not null, then this.user = window.user.

It's that simple! Sneak over and refresh your browser. And... our JavaScript instantly knows we're logged in. If we log out... yep! Our app doesn't explode. Woo! Yea... this is kinda fun!

More Complex Authentication

Ok! Authentication... including logging out and the frontend is done! If you're feeling great about this approach for you app, awesome! But if you're screaming:

  Ryan! My app is more complex... I have multiple APIs that talk to each other... or... my API will be exposed publicly... or I have some other reason that prevents me from this simple session-based authentication!

Then... don't worry. I've already gotten so many questions & feedback from the start of this tutorial that we're planning to create a separate, part 2 of this security tutorial where we'll talk about other common API use-cases and the right way to authenticate within those. We'll also talk about OAuth.

But in general, if you need a more custom authentication system - perhaps you can't use json_login because your login is more complex than just handling an email & password... or you're already planning some sort of token-based authentication - you can build that system by creating a custom Guard authenticator, which is something we talk about in our security tutorial.

Next, before we dive deep into denying and granting access to our API for different users, we need to talk about CSRF attacks and SameSite cookies. It turns out, if you use cookie-based authentication like we are, you may be vulnerable to CSRF attacks. Fortunately, there's a new, beautiful way to mitigate that.